+++ resources/config.yml
@@ -1,4 +1,4 @@
-#Choose if you want the plugin to send every commands to your discord server with a webhook (on/off)
+#Choose if the plugin should send the commands to your discord server using a webhook (on/off)
 webhook: off
 #enter the url of the webhook
 url:
