+++ src/SocialSpy/sqrrl/Main.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace SocialSpy\superbobby2000;
+namespace SocialSpy\sqrrl;
 
+use pocketmine\player\Player;
 use pocketmine\plugin\PluginBase;
 use pocketmine\command\CommandSender;
 use pocketmine\command\Command;
-use pocketmine\Player;
 use pocketmine\utils\TextFormat as C;
 
 use function in_array;
@@ -16,11 +16,11 @@ class Main extends PluginBase {
 
     public const PREFIX = "§9SocialSpy §6»§r ";
 
-    public static $SocialSpy = [];
+    public static array $SocialSpy = [];
 
-    protected static $main;
+    protected static self $main;
 
-    public function onEnable() {
+    public function onEnable(): void {
         @mkdir($this->getDataFolder());
         $this->saveDefaultConfig();
         self::$main = $this;
